requirements.txt
asgiref==3.3.1
astpretty==2.0.0
atomicwrites==1.4.0
attrs==20.3.0
cognitive-complexity==1.2.0
colorama==0.4.4
Django==3.1.3
django-cleanup==5.1.0
django-crispy-forms==1.10.0
django-phonenumber-field==5.0.0
eradicate==2.0.0
flake8==3.8.4
flake8-blind-except==0.1.1
flake8-builtins==1.5.3
flake8-cognitive-complexity==0.1.0
flake8-commas==2.0.0
flake8-comprehensions==3.3.0
flake8-django==1.1.1
flake8-eradicate==1.0.0
flake8-expression-complexity==0.0.8
flake8-functions==0.0.4
flake8-logging-format==0.6.0
flake8-polyfill==1.0.2
flake8-print==3.1.4
flake8-variables-names==0.0.3
iniconfig==1.1.1
mccabe==0.6.1
packaging==20.4
pep8-naming==0.11.1
phonenumbers==8.12.13
Pillow==8.0.1
pluggy==0.13.1
py==1.9.0
pycodestyle==2.6.0
pyflakes==2.2.0
pyparsing==2.4.7
pytest==6.1.2
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
